+++ src/components/Footer.tsx
@@ -107,11 +107,11 @@ const DownloadSimplyECallout = () => (
   <div sx={{ maxWidth: 300, flex: "0 1 auto", mt: 5 }}>
     <H3 sx={{ mt: 0, display: "flex", alignItems: "center" }}>
       <SvgPhone sx={{ mr: 1 }} />
-      Download SimplyE
+      Download Palace
     </H3>
     <Text>
       Our mobile app lets you browse, borrow and read from our whole collection
-      of eBooks and Audiobooks right on your phone!
+      of ebooks and audiobooks right on your phone!
     </Text>
     <div sx={{ width: "75%", overflow: "hidden", ml: -3 }}>
       <IosBadge sx={{ p: 3, pb: 0 }} />