+++ src/components/storeBadges/GooglePlayBadge.tsx
@@ -7,8 +7,9 @@ const GooglePlayBadge = (props: React.ComponentProps<"a">) => {
     <a
       rel="noopener noreferrer"
       target="__blank"
-      aria-label="Get SimplyE on the Google Play Store"
-      href="https://play.google.com/store/apps/details?id=org.nypl.simplified.simplye&pcampaignid=pcampaignidMKT-Other-global-all-co-prtnr-py-PartBadge-Mar2515-1"
+      aria-label="Get Palace on the Google Play Store"
+      // TODO: Replace with correct URL when Palace is in the Play Store.
+      href="https://play.google.com/store/apps/details?id=org.thepalaceproject.palace&pcampaignid=pcampaignidMKT-Other-global-all-co-prtnr-py-PartBadge-Mar2515-1"
       sx={{ display: "block" }}
       {...props}
     >